API.CORSMiddleware

Description

Based on @https://stackoverflow.com/questions/41399055/haskell-yesod-cors-problems-with-browsers-options-requests-when-doing-post-req Enables CORS requests to be made to this server by accepting all origins

Documentation

allowCsrf :: Middleware

x-csrf-token allowance. The following header will be set: Access-Control-Allow-Headers: x-csrf-token.

corsified :: Middleware

CORS middleware configured with appCorsResourcePolicy.

appCorsResourcePolicy :: CorsResourcePolicy

Cors resource policy to be used with corsified middleware.

This policy will set the following:

RequestHeaders: Content-Type

MethodsAllowed: OPTIONS, GET, PUT, POST