API.JSONData

Description

Documentation

data SpielRead

Representation of a request to read a BoGL file

Constructors

SpielRead

Fields

path :: String

data SpielShare

Representation of a request to share a prelude & gamefile

Constructors

SpielShare

Fields

preludeContent :: String

gameContent :: String

data SpielFile

Representation of a file that will be saved by the user

Constructors

SpielFile

Fields

fileName :: String

content :: String

data SpielCommand

Representation of input to the repl, from the user

Constructors

SpielCommand

Fields

prelude :: String

file :: String

input :: String

buffer :: [Val]

type Message = String

Error message

type LineNum = Int

Error line number

type ColNum = Int

Error column number

type FileName = String

Error file name

type BufferedBoards = [Val]

A buffer of boards which should be printed with the value of executing the expression

type ExecutionValue = Val

An execution value

data SpielResponse

Represents possible response categories from the server These are then parsed accordingly on the front-end

Constructors

SpielPrompt [Val]
represents a prompt for input and buffered boards

SpielSuccess String
represents a successful operation

SpielLoadResult String String
represents a successful load 1st prelude, 2nd gamefile

SpielGameResult String
represents a win/lose result

SpielTypeError TypeError
represents a type error

SpielParseError LineNum ColNum FileName Message
represents a parse error

SpielValue BufferedBoards ExecutionValue
represents a runtime error in spiel

SpielTypeHole LineNum ColNum Xtype
represents a typed hole that can be filled

SpielError String
fallback standard error, something went wrong in spiel

SpielRuntimeError String
generic runtime error

SpielTypes [(String, Type)]
List of typechecked types

Log String
String

type SpielResponses = [SpielResponse]

List of spiel responses

Orphan instances

FromJSON Val

Instance details

Methods

parseJSON :: Value -> Parser Val

parseJSONList :: Value -> Parser [Val]