API.RunCode

Description

Documentation

handleRunCode :: SpielCommand -> Handler SpielResponses

Runs literal code, lifting it into Handler Runs the contents of the prelude and gamefile without writing them out into files Once these both succeed, the new temporary files are parsed , and the response is returned