API.Server

Description

Handles starting up an instance of Servant Returns encoded board data on request, and computes boards when input data is taken

Documentation

startServer :: Int -> IO ()

Starts the language server

serverApp :: Application

Prepares a server application

type SpielApi = ("share" :> (ReqBody '[JSON] SpielShare :> Post '[JSON] SpielResponses)) :<|> (("load" :> (ReqBody '[JSON] SpielRead :> Post '[JSON] SpielResponse)) :<|> (("runCode" :> (ReqBody '[JSON] SpielCommand :> Post '[JSON] SpielResponses)) :<|> (("read" :> (ReqBody '[JSON] SpielRead :> Post '[JSON] SpielFile)) :<|> ("test" :> Get '[JSON] SpielResponses))))

Type of the API (Headers '[Header "Access-Control-Allow-Origin" String] SpielResponses)