Language.Syntax

Description

Spiel language AST. This slightly deviates from the specified syntax in spots.

Documentation

type Name = String

Names of games, signatures, etc.

data Game a

Game datatype

Constructors

Game

Fields

name :: Name

The name of the game

board :: BoardDef

Size and type of the board

input :: InputDef

Type of input

defns :: [ValDef a]

List of value definitions

data Signature

Signatures are a product of name and type.

Constructors

Sig Name Type

data Parlist

Parameters are lists of Name

Constructors

Pars [Name]

data ValDef a

Top level values are signatures paired with either an ordinary Equation or a list of Board Equations

Constructors

Val Signature (Equation a) a
Regular Value

BVal Signature [BoardEq a] a
Board value

data Equation a

Equations

Constructors

Veq Name (Expr a)
Value equations (a mapping from Name to Expr)

Feq Name Parlist (Expr a)
Function equations

data BoardEq a

Board equations are used to set positions on the board to an expression

Constructors

PosDef

Fields

boardEqName :: Name

xpos :: Pos

ypos :: Pos

boardExpr :: Expr a

data Pos

Types of individual positions for the x and y in a board equation

Constructors

Index Int
Singular index as

ForAll Name
All indices with a given name

data Expr a

Expressions

Constructors

I Int
Integer

S Name
Symbol

B Bool
Boolean

Ref Name
Reference to a variable

Tuple [Expr a]
Tuple of Expr

App Name (Expr a)
Function application with argument tuple

Binop Op (Expr a) (Expr a)
Binary operation

Let Name (Expr a) (Expr a)
Let binding

While (Expr a) (Expr a) [Name] (Expr a)
While loop

If (Expr a) (Expr a) (Expr a)
Conditional

Annotation a (Expr a)
Parameterized by the type of an annotation

HE Name
Type hole

data Op

Binary operations

Constructors

Plus
Addition (+)

Minus
Subtraction (-)

Times
Multiplication (*)

Div
Division (/)

Mod
Modulus (%)

Less
Less than comparison (<)

Leq
Less than equal to comparison (<=)

Equiv
Equivalent comparison (==)

NotEquiv
Not equivalent comparison (/=)

Geq
Greater than equal to comparison (>=)

Greater
Greater than comparison (>)

Get
Gets contents from a position on a board (!)

arithmetic :: Op -> Bool

Check if a binary operator is arithmetic

relational :: Op -> Bool

Check if a binary operator is relational (excluding == and /=)

equiv :: Op -> Bool

Check if a binary operator is == or /=

deAnnotate :: Expr a -> Expr a

Deannotate an expression

clearAnn :: Expr a -> Expr ()

Clear the annotation from an expression

ident :: ValDef a -> Name

Get the identifier from a ValDef