Language.Types

Description

Documentation

data BoardDef

Board definition

Constructors

BoardDef

Fields

size :: (Int, Int)

size of the board

piece :: Xtype

types of pieces that can be on the board

data InputDef

Input definition

Constructors

InputDef

Fields

inputType :: Xtype

Type of input, Player inputs must be an accepted type

data Btype

Atomic types

Constructors

Booltype
Boolean

Itype
Int

AnySymbol
this is the type all symbols live in

Input
The input type specified at the top of the program

Board
A game board

Top
Really this is bottom FIXME

Undef
Not definable by a user (only occurs when typechecking)

data Xtype

Xtypes are sum types (or tuples of sum types) but restricted by the semantics to only contain Symbols after the atomic type. Note: ttypes are subsumed by xtypes in our implementation

Constructors

X Btype (Set String)

Tup [Xtype]

Hole String

data Ftype

A function type can be from a plain type to a plain type (no curried functions)

Constructors

Ft Xtype Xtype

data Type

A type is either a plain type or a function.

Constructors

Plain Xtype

Function Ftype

bnestx :: Btype -> Xtype

Nest a Btype as an Xtype

boolxt :: Xtype

Xtype smart constructor for Booltype

intxt :: Xtype

Xtype smart constructor for Itype

boardt :: Type

Type smart constructor for Board

boardxt :: Xtype

Xtype smart constructor for Board

p :: Btype -> Type

Nest a Btype as a Type