Parser.Error

Description

Holds some functions for reporting parser errors

Documentation

quote :: String -> String

Double quotes a string

quoteMany :: [String] -> String

Uses a list of strings to produce a single string of the double quoted strings

errRepeatParam :: [String] -> String -> String

Constructs an error message indicating which params were repeated in what context