Parser.Parser

Description

Documentation

parseLine :: String -> Either ParseError (Expr SourcePos)

Parse a single line as an expression

parsePreludeFromText :: String -> Either ParseError ([Maybe (ValDef SourcePos)], ParState)

Parse the prelude from text

parseGameFromText :: String -> ([Maybe (ValDef SourcePos)], ParState) -> ParseResult

Parse a game from text and the result of a previous parse (e.g. the prelude) Such as in the case of the function above parsePreludeFromtext

parseGameFile :: String -> IO ParseResult

Parse a game from a string

parsePreludeAndGameText :: String -> String -> IO ParseResult

Parse a prelude and game from text directly, without a file

expr :: Parser (Expr SourcePos)

Parser for the Expr datatype

isLeft :: Either a b -> Bool

Return True if the given value is a Left-value, False otherwise.

Examples

Expand

Basic usage:

>>> isLeft (Left "foo")
True
>>> isLeft (Right 3)
False

Assuming a Left value signifies some sort of error, we can use isLeft to write a very simple error-reporting function that does absolutely nothing in the case of success, and outputs "ERROR" if any error occurred.

This example shows how isLeft might be used to avoid pattern matching when one does not care about the value contained in the constructor:

>>> import Control.Monad ( when )
>>> let report e = when (isLeft e) $ putStrLn "ERROR"
>>> report (Right 1)
>>> report (Left "parse error")
ERROR

Since: base-4.7.0.0

parseAll :: Parser a -> String -> String -> Either ParseError a

Uses the parser p to parse all input, throws an error if anything is left over

valdef :: Parser (ValDef SourcePos)

Value definitions

xtype :: Parser Xtype

Extended types: types after the first are restricted to symbols

boardeqn :: String -> Parser (BoardEq SourcePos)

Board equations

equation :: Parser (Equation SourcePos)

Equations

decl :: Parser (Maybe (ValDef SourcePos))

Parse a typesyn or a valdef

parseGame :: [ValDef SourcePos] -> Parser (Game SourcePos)

Game definition

typesyn :: Parser ()

Type Synonym Definition, but returns nothing

type Parser = Parsec String ParState

Parser type