Runtime.Builtins

Description

Documentation

single :: Xtype -> Xtype

Produces a Tuple type wrapping a single Xtype

builtinT :: Xtype -> Xtype -> [(String, Type)]

List of builtin function signatures

place :: [Val] -> Eval Val

places a piece on a board and also adds this new board to the display buffer. We only want the latest version of each unique board, so filter out its predecessor.

builtinsChecker :: [Char] -> [Val] -> Eval Val

Verifies the parameters are of the expected number & type from their respective lambdas A final case reports a Runtime error in case anything slips by, which indicates that the number and/or type of parameters didn't match what was expected for that builtin Note: This prevents a crash and gives somewhat better feedback to users, but this should not normally be reachable if the typechecker verifies the expressions correctly in advance

builtins :: [(Name, [Val] -> Eval Val)]

List of builtin functions, with definitions for evaluation

builtinRefs :: [(Name, Eval Val)]

Builtin references, just input for now

type Count = (Int, Int, Int, Int)

the count of adjacent cells in four directions (above, diagonal 10:30, left, diagonal 7:30)

type CountMap = Map (Int, Int) Count

Map of the count

peek :: (Int, Int) -> CountMap -> Count

A safe map lookup function which returns a default value for keys not in the map

addCell :: (Int, Int) -> CountMap -> CountMap

Adds a cell into the count map

checkCell :: Val -> ((Int, Int), Val) -> CountMap -> CountMap

Checks a cell in the count map

checkCells :: Board -> Val -> [Int]

Scans cells downwards by column (the order given by (assocs b) with (x,y) coords) each cell's count is the increment of the counts of the four cells before it

countCol :: Board -> Val -> Int

Counts cols, rows, and diagonals in a board

countRow :: Board -> Val -> Int

Counts cols, rows, and diagonals in a board

countDiag :: Board -> Val -> Int

Counts cols, rows, and diagonals in a board

inARow :: Board -> Val -> Int -> Bool

checks whether a board has i cells containing v in a row