Runtime.Eval

Description

Documentation

bindings :: (Int, Int) -> [ValDef a] -> Writer [Exception] Env

Produce all of the bindings from a list of value definitions. This is done sequentially. Errors are reported as they're found.

bindings_ :: (Int, Int) -> [ValDef a] -> Env

Partial application of of args to bindings, allows passing in just a list of ValDefs

bind :: ValDef a -> (Name, Eval Val)

Binds a value equation

Binds a ValDef to a name and a Val in the Eval monad

updateBoard :: Board -> (Int, Int) -> BoardEq a -> Val -> Board

Updates a board

posMatches :: Pos -> Pos -> (Int, Int) -> Bool

Check if a Pos matches a coordinate pair

tryUnsafeBoardAccess :: (Int, Int) -> Board -> Val

Attempt to access a position on the board that may be invalid If the position is invalid, return an Err value instead of causing an actual array access error

evalBinOp :: Op -> Expr a -> Expr a -> Eval Val

Binary operation evaluation

evalEq :: (Val -> Val -> Bool) -> Expr a -> Expr a -> Eval Val

evaluates the == and /= operations

evalCompareOpInt :: (Int -> Int -> Bool) -> Expr a -> Expr a -> Eval Val

evaluates comparison operations for only Ints (except for == & /=)

evalNumOp :: String -> (Int -> Int -> Int) -> Expr a -> Expr a -> Eval Val

evaluates numerical operations

eval :: Expr a -> Eval Val

Evaluates an expression at runtime, producing a Val in the Eval monad

runWithBuffer :: Env -> Buffer -> Expr SourcePos -> Either ([Val], Buffer) ([Val], Val)

Runs an expression under a given environment and a given buffer, producing a result of either a list of values and a buffer, or a list of values and a single value