Runtime.Monad

Description

Evaluation Monad

Documentation

type Eval a = StateT Buffer (ExceptT Exception (ReaderT Env Identity)) a

Eval Monad transformer

data Env

Call-by-value semantics

Constructors

Env

Fields

evalEnv :: EvalEnv

boardSize :: (Int, Int)

getBounds :: Eval (Int, Int)

Uses the StateT monad to request the typed board dimensions in the runtime environment

getEnv :: Eval EvalEnv

Uses the StateT monad to get the evaluation environment in the runtime environment

emptyEnv :: (Int, Int) -> Env

Produces an empty environment for testing, and for starting evaluations

modifyEval :: (EvalEnv -> EvalEnv) -> Env -> Env

Modifies the evaluation environment, producing a new environment

type Buffer = ([Val], [Val], Int)

Input buffer and display buffer. The display buffer stores all boards which are to be printed on the front end after a board is updated Additionally counts the number of evaluation iterations, stopping after a fixed amount with a 'Error $ "Stack Overflow!"'

data Exception

Exceptions

Constructors

NeedInput [Val]
Ran out of input and here's the buffered display boards

Error String
Encountered a runtime error

evalWithLimit :: Eval Val -> Eval Val

Take an expressions, and before evaluating it checks & updates evaluation iterations If the count is less than the limit, continue evaluating, otherwise return an error instead of evluating further. Prevents infinite loops via recursion, while, or self referencing value equations, among other things

runEval :: Env -> Buffer -> Eval a -> Either Exception a

Evaluation occurs in the Identity monad with these side effects: ReaderT: Evaluation enviroment, board size and piece type, and input type StateT: Input buffer, used for reading input

Evaluate in the environment given, with a buffer.

extScope :: EvalEnv -> Eval a -> Eval a

Evaluate with an extended scope

lookupName :: Name -> Eval (Maybe Val)

Lookup a name in the environment FIXME

waitForInput :: [Val] -> Eval a

Ask for input, displaying a value to the user

err :: String -> Eval a

Converts a string into an evaluation error

readTape :: Eval Val

Read input

unpackBool :: Val -> Maybe Bool

Helper function to get the Bool out of a value. This is a partial function.