Typechecker.Monad

Description

Documentation

type TypeEnv = [(Name, Type)]

Types in the environment

data Env

Typechecker environment

Constructors

Env

Fields

types :: TypeEnv

input :: Xtype

piece :: Xtype

size :: (Int, Int)

initEnv :: Xtype -> Xtype -> (Int, Int) -> Env

Initial empty environment

exampleEnv :: Env

An example environment for interal use (e.g. testing, ghci)

data Stat

Typechecker state

Constructors

Stat

Fields

holes :: TypeEnv

source :: Maybe (Expr SourcePos)

pos :: SourcePos

type Typechecked a = StateT Stat (ReaderT Env (ExceptT TypeError Identity)) a

Typechecking monad

typecheck :: Env -> Typechecked a -> Either TypeError (a, Stat)

Run a computation inside of the typechecking monad

typeHoles :: Env -> Typechecked a -> Either TypeError (a, TypeEnv)

Typecheck type holes

extendEnv :: Env -> (Name, Type) -> Env

Add some types to the environment

getEnv :: Typechecked TypeEnv

Get the type environment

getInput :: Typechecked Xtype

Get the input type

getPiece :: Typechecked Xtype

Get the piece type

getSize :: Typechecked (Int, Int)

Get the board size

inBounds :: (Int, Int) -> Typechecked Bool

Check whether (x,y) is in the bounds of the board

localEnv :: ([(Name, Type)] -> [(Name, Type)]) -> Typechecked a -> Typechecked a

Extend the environment

getHoles :: Typechecked TypeEnv

Get the current type holes

setSrc :: Expr SourcePos -> Typechecked ()

Set the source line

setPos :: SourcePos -> Typechecked ()

Set the position

getPos :: Typechecked SourcePos

Get the position

getSrc :: Typechecked (Expr SourcePos)

Get the source line

getType :: Name -> Typechecked Type

Get a type from the environment

addHole :: (Name, Type) -> Typechecked ()

add a type hole

unify :: Xtype -> Xtype -> Typechecked Xtype

Attempt to unify two types

t :: Btype -> Typechecked Xtype

Returns a typechecked base type

data TypeError

Encoding the different type errors as types should let us do interesting things with them

Constructors

Mismatch
Couldn't match two types in an expression

Fields

t1 :: Type

t2 :: Type

srcPos2 :: Expr SourcePos

srcPos :: SourcePos

NotBound
Name isn't (yet) bound in the enviroment

Fields

name :: Name

srcPos :: SourcePos

SigMismatch
couldn't match the type of an equation with its signature

Fields

name :: Name

sigType :: Type

actualType :: Type

srcPos :: SourcePos

Unknown
Errors that "shouldn't happen"

Fields

msg :: String

srcPos :: SourcePos

BadOp
Can't perform a primitive operation

Fields

op :: Op

t1 :: Type

t2 :: Type

srcPos2 :: Expr SourcePos

srcPos :: SourcePos

OutOfBounds

Fields

xpos :: Pos

ypos :: Pos

srcPos :: SourcePos

BadApp
An attempt to apply a non-function expr as if it were a function

Fields

name :: Name

arg :: Expr SourcePos

srcPos :: SourcePos

Dereff
An attempt to dereference a function

Fields

name :: Name

typ :: Type

srcPos :: SourcePos

Uninitialized

Fields

name :: Name

srcPos :: SourcePos

mismatch :: Type -> Type -> Typechecked a

Type mismatch error

notbound :: Name -> Typechecked a

Not bound type error

sigmismatch :: Name -> Type -> Type -> Typechecked a

Signature mismatch type error

unknown :: String -> Typechecked a

Unknown type error

badop :: Op -> Type -> Type -> Typechecked a

Bad Op type error

outofbounds :: Pos -> Pos -> Typechecked a

Out of Bounds type error

uninitialized :: Name -> Typechecked a

Uninitialized board type error

badapp :: Name -> Expr SourcePos -> Typechecked a

Bad function application type error

dereff :: Name -> Type -> Typechecked a

Cannot dereference function type error

extensions :: Xtype -> Typechecked (Set Name)

Retrieve the extensions from an Xtype

errString :: SourcePos -> String

Produce a human readable error string from a source position