Typechecker.Typechecker

Description

Documentation

tcexpr :: Env -> Expr SourcePos -> Either TypeError (Xtype, TypeEnv)

Run the typechecker on an Expr and report any errors to the console.

environment :: BoardDef -> InputDef -> [ValDef SourcePos] -> Env

Produce the environment

runTypeCheck :: BoardDef -> InputDef -> [ValDef SourcePos] -> Writer [Either (ValDef SourcePos, TypeError) (Name, Type)] Env

Runs the typechecker on a board def, input def, list of valdefs, and produces results of errors or successfully typechecked names recursion is not allowed by this.

tc :: Game SourcePos -> TcResult

Typecheck a game and produce a typechecker result

printT :: (Xtype, TypeEnv) -> String

Produce a string of types in the environment

data TcResult

Typechecker Result

Constructors

Tc

Fields

success :: Bool

Success result

e :: Env

Env associated with the result

errors :: [(ValDef SourcePos, TypeError)]

Typechecker errors

rtypes :: [(Name, Type)]

List of (name,type) pairs

showTCError :: (ValDef SourcePos, TypeError) -> String

Produces an error corresponding to a typechecker error